feat: enhance booking validation and error handling, update payment processing logic, and localize error messages

+++ app/lib/features/client/bookings/bloc/client_booking_details_event.dart
@@ -21,7 +21,7 @@ final class ClientBookingSubmitPressed extends ClientBookingDetailsEvent {
   final int seatsCount;
   final String name;
   final String contact;
-  final int fallbackPaymentAmount;
+  final double fallbackPaymentAmount;
   final String itemName;
 
   @override
